refactor: rename vectordbinterface.py to base.py for better clarity
