+++ 09-VectorStore/utils/neo4j_interface.py
@@ -1,5 +1,5 @@
 import neo4j
-from .vectordbinterface import DocumentManager
+from .base import DocumentManager
 from langchain_core.documents import Document
 from typing import List, Union, Dict, Any, Optional, Iterable
 from concurrent.futures import ThreadPoolExecutor, as_completed
